requirements.txt
torch
torchvision
opencv-python
albumentations
tqdm
pyyaml